requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
anthropic>=0.42.0
openai>=1.60.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
Pillow>=10.0.0
google-genai>=1.0.0
httpx>=0.25.0
fpdf2>=2.8.0
